+++ apps/web/src/features/freee/components/freee-insert-form.tsx
@@ -5,6 +5,7 @@ import { useState } from "react";
 
 import { Button } from "../../../components/ui/button/v5";
 import { errorPopup, successPopup } from "../../../function/successPopup";
+import { submitFreeeDeals } from "../actions/freee-submit-actions";
 import type { UnifiedRecord } from "../types/unified-record";
 
 /**
@@ -28,105 +29,26 @@ export const FreeeInsertForm: FC<{ initialRecords?: UnifiedRecord[] }> = ({
     setRecords(newRecords);
   };
 
+  // メッセージ定義
+  const successMessage = "freeeへのデータ送信が完了しました";
+  const errorMessage = "データの送信に失敗しました。もう一度お試しください。";
+
   // フォーム送信ハンドラ
   const handleSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
 
-    // レコードを処理して、APIに送信するデータ形式に変換
-    const processedRecords = records.map((record) => {
-      // 基本情報は各レコードで同じなので、最初のレコードから取得
-      const formData = {
-        issue_date: record.issue_date,
-        type: record.type,
-        company_id: record.company_id
-          ? Number.parseInt(record.company_id, 10)
-          : null,
-        due_date: record.due_date,
-        partner_id: record.partner_id
-          ? Number.parseInt(record.partner_id, 10)
-          : null,
-        partner_code: record.partner_code,
-        ref_number: record.ref_number,
-      };
-
-      // 明細情報
-      const detail = {
-        tax_code: record.tax_code ? Number.parseInt(record.tax_code, 10) : null,
-        account_item_id: record.account_item_id
-          ? Number.parseInt(record.account_item_id, 10)
-          : null,
-        amount: record.amount ? Number.parseInt(record.amount, 10) : null,
-        item_id: record.item_id ? Number.parseInt(record.item_id, 10) : null,
-        section_id: record.section_id
-          ? Number.parseInt(record.section_id, 10)
-          : null,
-        tag_ids: record.tag_ids
-          .map((id) => (id ? Number.parseInt(id, 10) : null))
-          .filter(Boolean),
-        description: record.description,
-        vat: record.vat ? Number.parseInt(record.vat, 10) : null,
-      };
-
-      // 支払情報
-      const payment = {
-        amount: record.payment_amount
-          ? Number.parseInt(record.payment_amount, 10)
-          : null,
-        from_walletable_id: record.from_walletable_id
-          ? Number.parseInt(record.from_walletable_id, 10)
-          : null,
-        from_walletable_type: record.from_walletable_type,
-        date: record.payment_date,
-      };
-
-      // 領収書ID
-      const receiptId = record.receipt_id
-        ? Number.parseInt(record.receipt_id, 10)
-        : null;
-
-      return {
-        formData,
-        detail,
-        payment,
-        receiptId,
-      };
-    });
-
-    // APIに送信するデータ形式に変換
-    const requestData = {
-      ...processedRecords[0].formData, // 基本情報は最初のレコードから取得
-      details: processedRecords.map((record) => record.detail),
-      payments: processedRecords.map((record) => record.payment),
-      receipt_ids: processedRecords
-        .map((record) => record.receiptId)
-        .filter(Boolean),
-    };
-
     try {
-      // ここでAPIリクエストを送信
-      console.log("送信データ:", requestData);
+      // サーバーアクションを呼び出してデータを送信
+      const result = await submitFreeeDeals(records);
 
-      // TODO: 実際のAPIリクエストを実装
-      // const response = await fetch('/api/freee/deals', {
-      //   method: 'POST',
-      //   headers: {
-      //     'Content-Type': 'application/json',
-      //   },
-      //   body: JSON.stringify(requestData),
-      // });
-
-      // if (!response.ok) {
-      //   throw new Error('APIリクエストが失敗しました');
-      // }
-
-      // const result = await response.json();
-      // console.log('API応答:', result);
-
-      // 成功メッセージを表示
-      successPopup("freeeへのデータ送信が完了しました");
+      if (result.success) {
+        successPopup(successMessage);
+      } else {
+        errorPopup(result.error?.message || errorMessage);
+      }
     } catch (error) {
       console.error("Error submitting data:", error);
-      errorPopup("データの送信に失敗しました。もう一度お試しください。");
+      errorPopup(errorMessage);
     }
   };
